refactor: feature/#50 UserService to UserServiceImpl
